requirements.txt
fastapi
uvicorn
PyMuPDF
Pillow
numpy
opencv-python-headless
easyocr
openai
python-dotenv
python-multipart
jinja2
